+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### New Features
 
+- Introduce the filter feature in the file browser.
 - Introduce the “Insert comment delimiters after indent” option to the Edit settings pane.
 - Revert the "Append a space to comment delimiter" option to the Edit settings pane.